+++ src/api/admin/__init__.py
@@ -2,5 +2,6 @@ from .base import *  # noqa: F401, F403
 from .classification import ImageAdmin  # noqa: F401, F403
 from .gfcr import GFCRIndicatorSetAdmin  # noqa: F401, F403
 from .mermaid import *  # noqa: F401, F403
+from .protocols import *  # noqa: F401, F403
 from .revisions import RevisionAdmin  # noqa: F401, F403
 from .summaries import SummaryCacheQueueAdmin  # noqa: F401, F403
