+++ src/api/admin/protocols/__init__.py
@@ -1,0 +1,4 @@
+from .beltfish import *  # noqa: F401, F403
+from .benthic import *  # noqa: F401, F403
+from .bleachingqc import *  # noqa: F401, F403
+from .macroinvertebrate import *  # noqa: F401, F403
