+++ src/api/scripts/benthicattrib_update_1902.py
@@ -1,4 +1,4 @@
-from api.models.mermaid import BenthicAttribute
+from api.models import BenthicAttribute
 
 
 class BenthicAttributeUpdateGenerator(object):
