+++ src/api/scripts/load_test_data.py
@@ -1,5 +1,4 @@
-from api.models.base import AuthUser, Profile
-from api.models.mermaid import (
+from api.models import (
     Country,
     Management,
     ManagementParty,
@@ -10,6 +9,7 @@ from api.models.mermaid import (
     ReefZone,
     Site,
 )
+from api.models.base import AuthUser, Profile
 
 
 class TestDataGenerator(object):
